Factor out query code that we will reuse in get_medication_date_matrix
